Finished the base of the game.
A simple game done. The snake incrases is size when eating the food that is randomly put on the screen.
If the snake touches the boards or touch himself the game ends and a message apears saying that you lost.
Now its time to add some things to make the game prettier and better.